Add files via upload
To help you get started, Chapter 1 of this handbook identifies some of the primary reasons that people are not successful on Institutes exams. Other chapters focus on ways to overcome exam-related problems. Try the techniques and suggestions in this handbook to find the ones that work best for you. After some orientation, you will develop a routine that both fits your schedule and helps you succeed.